requirements.txt
beautifulsoup4==4.12.3
Eel==0.16.0
GitPython==3.1.43
GitPython==3.1.43
Requests==2.32.3
